+++ train.py
@@ -58,8 +58,8 @@ gradient_accumulation_steps = config.gradient_accumulation_steps
 
 # poor man's data loader
 data_dir = os.path.join(config.data_dir, config.dataset)
-train_data = np.memmap(os.path.join(data_dir, 'train.bin'), dtype=np.uint32, mode='r')
-val_data = np.memmap(os.path.join(data_dir, 'val.bin'), dtype=np.uint32, mode='r')
+train_data = np.memmap(os.path.join(data_dir, 'train.bin'), dtype=np.uint16, mode='r')
+val_data = np.memmap(os.path.join(data_dir, 'val.bin'), dtype=np.uint16, mode='r')
 dataset_train_x_start = config.dataset_seq_train_start if config.dataset_seq_train_start is not None else random.randint(0, batch_size-1)
 def get_batch(split):
     data = train_data if split == 'train' else val_data
